A string of cleanup and refactors
