+++ androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/base/data/persistence/GamesDatabase.kt
@@ -1,11 +1,11 @@
-package com.lindenlabs.scorebook.androidApp.data.persistence
+package com.lindenlabs.scorebook.androidApp.base.data.persistence
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Game
+import com.lindenlabs.scorebook.androidApp.base.data.raw.Game
 
 @Database(entities = [Game::class], version = 1, exportSchema = false)
 @TypeConverters(Converters.UUIDConverter::class, Converters.PlayerListConverter::class,  Converters.PlayerConverter::class)
