+++ androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/base/data/raw/Game.kt
@@ -1,13 +1,10 @@
-package com.lindenlabs.scorebook.androidApp.screens.home.data.model
+package com.lindenlabs.scorebook.androidApp.base.data.raw
 
 import android.os.Parcelable
 import androidx.room.*
-import com.lindenlabs.scorebook.androidApp.data.persistence.Converters
-import com.lindenlabs.scorebook.androidApp.data.persistence.Converters.*
-import com.lindenlabs.scorebook.androidApp.screens.home.presentation.GameStrategy
+import com.lindenlabs.scorebook.androidApp.base.data.persistence.Converters.*
+import com.lindenlabs.scorebook.androidApp.screens.home.presentation.entities.GameStrategy
 import kotlinx.android.parcel.Parcelize
-import kotlinx.android.parcel.RawValue
-import java.lang.IllegalStateException
 import java.util.*
 
 typealias StalematePair = Pair<Player, Player>
