--- androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/data/PlayerRepository.kt
@@ -1,11 +1,0 @@
-package com.lindenlabs.scorebook.androidApp.data
-
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Player
-
-class PlayerRepository : PlayerDataSource {
-    override val players: MutableList<Player> = mutableListOf()
-
-    override fun addPlayer(player: Player) {
-        players += player
-    }
-}
