--- androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/navigation/FragmentTarget.kt
@@ -1,6 +1,0 @@
-package com.lindenlabs.scorebook.androidApp.navigation
-
-import android.os.Bundle
-import androidx.annotation.IdRes
-
-data class FragmentTarget(@IdRes val targetId: Int, val bundle: Bundle)
