+++ androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/screens/addplayers/AddPlayersViewEvent.kt
@@ -1,6 +1,6 @@
 package com.lindenlabs.scorebook.androidApp.screens.addplayers
 
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Game
+import com.lindenlabs.scorebook.androidApp.base.data.raw.Game
 
 sealed class AddPlayersViewEvent {
     data class NavigateToGameDetail(val game: Game) : AddPlayersViewEvent()
