+++ androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/screens/home/domain/GetOpenGames.kt
@@ -1,6 +1,6 @@
 package com.lindenlabs.scorebook.androidApp.screens.home.domain
 
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Game
+import com.lindenlabs.scorebook.androidApp.base.data.raw.Game
 
 class GetOpenGames(private val games: List<Game>) {
     operator fun invoke() = games.filterNot { it.isClosed }
