--- androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/screens/home/presentation/GameStrategy.kt
@@ -1,6 +1,0 @@
-package com.lindenlabs.scorebook.androidApp.screens.home.presentation
-
-enum class GameStrategy {
-    HighestScoreWins,
-    LowestScoreWins
-}
