+++ androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/screens/home/presentation/HomeViewModel.kt
@@ -4,17 +4,16 @@ import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
-import com.lindenlabs.scorebook.androidApp.base.GameEngine
-import com.lindenlabs.scorebook.androidApp.data.GameDataSource
-import com.lindenlabs.scorebook.androidApp.data.PersistentGameRepository
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Game
-import com.lindenlabs.scorebook.androidApp.screens.home.presentation.GameStrategy.HighestScoreWins
-import com.lindenlabs.scorebook.androidApp.screens.home.presentation.GameStrategy.LowestScoreWins
+import com.lindenlabs.scorebook.androidApp.base.domain.GameEngine
+import com.lindenlabs.scorebook.androidApp.base.domain.PersistentGameRepository
+import com.lindenlabs.scorebook.androidApp.base.data.raw.Game
+import com.lindenlabs.scorebook.androidApp.base.domain.GameDataSource
+import com.lindenlabs.scorebook.androidApp.screens.home.presentation.entities.*
 import com.lindenlabs.scorebook.androidApp.screens.home.presentation.entities.GameInteraction
+import com.lindenlabs.scorebook.androidApp.screens.home.presentation.entities.GameStrategy.HighestScoreWins
+import com.lindenlabs.scorebook.androidApp.screens.home.presentation.entities.GameStrategy.LowestScoreWins
 import com.lindenlabs.scorebook.androidApp.screens.home.presentation.entities.GameInteraction.GameClicked
 import com.lindenlabs.scorebook.androidApp.screens.home.presentation.entities.GameInteraction.GameDetailsEntered
-import com.lindenlabs.scorebook.androidApp.screens.home.presentation.entities.GamesWrapper
-import com.lindenlabs.scorebook.androidApp.screens.home.presentation.entities.HomeViewEvent
 import com.lindenlabs.scorebook.androidApp.screens.home.presentation.entities.HomeViewEvent.*
 import com.lindenlabs.scorebook.androidApp.screens.home.presentation.entities.HomeViewState
 import kotlinx.coroutines.launch
