+++ androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/screens/home/presentation/entities/GameInteraction.kt
@@ -1,6 +1,6 @@
 package com.lindenlabs.scorebook.androidApp.screens.home.presentation.entities
 
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Game
+import com.lindenlabs.scorebook.androidApp.base.data.raw.Game
 
 internal sealed class GameInteraction {
     data class GameDetailsEntered(val name: String?, val lowestScoreWins: Boolean = false) : GameInteraction()
