+++ androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/screens/home/presentation/entities/GameStrategy.kt
@@ -1,0 +1,6 @@
+package com.lindenlabs.scorebook.androidApp.screens.home.presentation.entities
+
+enum class GameStrategy {
+    HighestScoreWins,
+    LowestScoreWins
+}
