+++ androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/screens/home/presentation/entities/GamesWrapper.kt
@@ -1,5 +1,5 @@
 package com.lindenlabs.scorebook.androidApp.screens.home.presentation.entities
 
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Game
+import com.lindenlabs.scorebook.androidApp.base.data.raw.Game
 
 data class GamesWrapper(val openGames: List<Game>, val closedGames: List<Game>)
