+++ androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/screens/home/presentation/entities/HomeViewEvent.kt
@@ -1,6 +1,6 @@
 package com.lindenlabs.scorebook.androidApp.screens.home.presentation.entities
 
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Game
+import com.lindenlabs.scorebook.androidApp.base.data.raw.Game
 
 sealed class HomeViewEvent {
     data class AlertNoTextEntered(val errorText: String = "Enter a game name to begin") : HomeViewEvent()
