+++ androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/screens/scorebookdetail/GameViewEntityMapper.kt
@@ -2,7 +2,7 @@ package com.lindenlabs.scorebook.androidApp.screens.scorebookdetail
 
 import com.lindenlabs.scorebook.androidApp.screens.scorebookdetail.entities.ScoreBookEntity
 import com.lindenlabs.scorebook.androidApp.screens.scorebookdetail.entities.ScoreBookInteraction
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Player
+import com.lindenlabs.scorebook.androidApp.base.data.raw.Player
 
 class GameViewEntityMapper {
