+++ androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/screens/scorebookdetail/entities/ScoreBookInteraction.kt
@@ -1,6 +1,6 @@
 package com.lindenlabs.scorebook.androidApp.screens.scorebookdetail.entities
 
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Player
+import com.lindenlabs.scorebook.androidApp.base.data.raw.Player
 
 sealed class ScoreBookInteraction {
     data class PlayerClicked(val player : Player) : ScoreBookInteraction()
