+++ androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/screens/scorebookdetail/entities/ScoreBookViewEvent.kt
@@ -1,7 +1,7 @@
 package com.lindenlabs.scorebook.androidApp.screens.scorebookdetail.entities
 
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Game
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Player
+import com.lindenlabs.scorebook.androidApp.base.data.raw.Game
+import com.lindenlabs.scorebook.androidApp.base.data.raw.Player
 
 sealed class ScoreBookViewEvent {
     data class AddPlayersClicked(val game: Game) : ScoreBookViewEvent()
