+++ androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/screens/updatepoints/UpdatePointsViewEvent.kt
@@ -1,7 +1,7 @@
 package com.lindenlabs.scorebook.androidApp.screens.updatepoints
 
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Game
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Player
+import com.lindenlabs.scorebook.androidApp.base.data.raw.Game
+import com.lindenlabs.scorebook.androidApp.base.data.raw.Player
 
 sealed class UpdatePointsViewEvent {
     data class ScoreUpdated(val player: Player, val game: Game) : UpdatePointsViewEvent()
