+++ androidApp/src/main/java/com/lindenlabs/scorebook/androidApp/screens/updatepoints/UpdatePointsViewState.kt
@@ -1,7 +1,7 @@
 package com.lindenlabs.scorebook.androidApp.screens.updatepoints
 
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Game
-import com.lindenlabs.scorebook.androidApp.screens.home.data.model.Player
+import com.lindenlabs.scorebook.androidApp.base.data.raw.Game
+import com.lindenlabs.scorebook.androidApp.base.data.raw.Player
 
 sealed class UpdatePointsViewState {
     data class ScreenOpened(val game: Game, val player: Player) : UpdatePointsViewState()
